refactor: split main routes into modules
