+++ server/assinaturaRoutes.ts
@@ -1,0 +1,127 @@
+import { Express, Request, Response } from 'express';
+import { db } from './db';
+import * as schema from '../shared/schema';
+
+export function registerAssinaturaRoutes(app: Express) {
+  // Criar cliente no Asaas
+  app.post('/api/asaas/criar-cliente', async (req: Request, res: Response) => {
+    try {
+      const { nome, email, telefone, cpf } = req.body;
+
+      if (!nome) {
+        return res.status(400).json({ message: 'Nome é obrigatório' });
+      }
+
+      const asaasApiKey = process.env.ASAAS_AND;
+      if (!asaasApiKey) {
+        return res.status(500).json({ message: 'Chave API ASAAS não configurada' });
+      }
+
+      const customerData = {
+        name: nome,
+        email: email || undefined,
+        phone: telefone || undefined,
+        cpfCnpj: cpf || undefined
+      };
+
+      const baseUrl = 'https://www.asaas.com/api/v3';
+      const response = await fetch(`${baseUrl}/customers`, {
+        method: 'POST',
+        headers: {
+          'access_token': asaasApiKey,
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify(customerData)
+      });
+
+      const result = await response.json();
+
+      if (response.ok) {
+        res.json({
+          success: true,
+          customer: result,
+          ambiente: 'PRODUÇÃO',
+          baseUrl: baseUrl,
+          message: 'Cliente criado com sucesso'
+        });
+      } else {
+        res.status(400).json({
+          success: false,
+          message: 'Erro ao criar cliente',
+          ambiente: 'PRODUÇÃO',
+          baseUrl: baseUrl,
+          error: result
+        });
+      }
+    } catch (error) {
+      console.error('Erro ao criar cliente:', error);
+      res.status(500).json({
+        success: false,
+        message: 'Erro interno do servidor',
+        error: error instanceof Error ? error.message : 'Erro desconhecido'
+      });
+    }
+  });
+
+  // Criar assinatura no Asaas
+  app.post('/api/asaas/criar-assinatura', async (req: Request, res: Response) => {
+    try {
+      const { customerId, value, description } = req.body;
+
+      if (!customerId || !value) {
+        return res.status(400).json({ message: 'customerId e value são obrigatórios' });
+      }
+
+      const asaasApiKey = process.env.ASAAS_AND;
+      if (!asaasApiKey) {
+        return res.status(500).json({ message: 'Chave API ASAAS não configurada' });
+      }
+
+      const subscriptionData = {
+        customer: customerId,
+        billingType: 'PIX',
+        nextDueDate: new Date().toISOString().split('T')[0],
+        value: value,
+        cycle: 'MONTHLY',
+        description: description || 'Assinatura Mensal'
+      };
+
+      const baseUrl = 'https://www.asaas.com/api/v3';
+      const response = await fetch(`${baseUrl}/subscriptions`, {
+        method: 'POST',
+        headers: {
+          'access_token': asaasApiKey,
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify(subscriptionData)
+      });
+
+      const result = await response.json();
+
+      if (response.ok) {
+        res.json({
+          success: true,
+          subscription: result,
+          ambiente: 'PRODUÇÃO',
+          baseUrl: baseUrl,
+          message: 'Assinatura criada com sucesso'
+        });
+      } else {
+        res.status(400).json({
+          success: false,
+          message: 'Erro ao criar assinatura',
+          ambiente: 'PRODUÇÃO',
+          baseUrl: baseUrl,
+          error: result
+        });
+      }
+    } catch (error) {
+      console.error('Erro ao criar assinatura:', error);
+      res.status(500).json({
+        success: false,
+        message: 'Erro interno do servidor',
+        error: error instanceof Error ? error.message : 'Erro desconhecido'
+      });
+    }
+  });
+}
