+++ server/routes.ts
@@ -5,6 +5,9 @@ import { db } from './db';
 import * as schema from '../shared/schema';
 import { eq, like, sql, and, gte, lte } from 'drizzle-orm';
 import bcrypt from 'bcrypt';
+import { registerAuthRoutes } from './authRoutes';
+import { registerClienteRoutes } from './clienteRoutes';
+import { registerAssinaturaRoutes } from './assinaturaRoutes';
 
 // Middleware de autorização baseado exclusivamente no campo role
 function requireRole(allowedRoles: string[]) {
@@ -216,535 +219,13 @@ export async function registerRoutes(app: Express): Promise<Express> {
       next();
     }
   });
+
+  // Registrar rotas divididas
+  registerAuthRoutes(app);
+  registerClienteRoutes(app);
+  registerAssinaturaRoutes(app);
   
   // Nova função para determinar status baseado no mês e lógica de pagamentos vencidos
-  function determinarStatusClientePorMes(cliente: any, mesFiltro: string): string {
-    // Regra: Cliente é ATIVO se está em dia ou cobrança ainda não venceu
-    // Regra: Cliente é ATRASADO apenas se pagamento venceu e não foi confirmado
-    
-    // Para simplificação inicial, vamos usar a lógica baseada em:
-    // - notificationDisabled = true -> provavelmente cancelado/inativo
-    // - Cliente muito recente (menos de 30 dias) -> ativo (ainda dentro do prazo)
-    // - Clientes antigos sem notificationDisabled -> ativos (presumindo em dia)
-    
-    if (cliente.notificationDisabled || cliente.deleted) {
-      return 'atrasado'; // Cliente com notificação desabilitada ou deletado = atrasado/cancelado
-    }
-    
-    const dataCreated = new Date(cliente.dateCreated);
-    const agora = new Date();
-    const diasCriacao = Math.floor((agora.getTime() - dataCreated.getTime()) / (1000 * 3600 * 24));
-    
-    // Clientes muito antigos (mais de 90 dias) sem atividade recente podem estar atrasados
-    if (diasCriacao > 90 && cliente.notificationDisabled) {
-      return 'atrasado';
-    }
-    
-    // Por padrão, considera ativo (em dia ou dentro do prazo)
-    return 'ativo';
-  }
-
-  // Função auxiliar para verificar status do cliente baseado em cobranças
-  async function verificarStatusCliente(cliente: any, apiKey: string): Promise<'ativo' | 'inadimplente'> {
-    try {
-      const response = await fetch(`https://www.asaas.com/api/v3/payments?customer=${cliente.id}&limit=50`, {
-        headers: {
-          'access_token': apiKey,
-          'Content-Type': 'application/json'
-        }
-      });
-
-      if (!response.ok) {
-        console.log(`Erro ao buscar cobranças do cliente ${cliente.id}`);
-        return 'ativo'; // Default para ativo em caso de erro
-      }
-
-      const data = await response.json();
-      const cobrancas = data.data || [];
-      
-      if (cobrancas.length === 0) {
-        return 'ativo'; // Cliente sem cobranças = ativo
-      }
-
-      // Ordenar por data de vencimento (mais recente primeiro)
-      cobrancas.sort((a: any, b: any) => new Date(b.dueDate).getTime() - new Date(a.dueDate).getTime());
-      
-      const hoje = new Date();
-      hoje.setHours(0, 0, 0, 0);
-
-      // Buscar a última cobrança paga (RECEIVED ou CONFIRMED)
-      const cobrancasPagas = cobrancas.filter((c: any) => c.status === 'RECEIVED' || c.status === 'CONFIRMED');
-      const ultimaCobrancaPaga = cobrancasPagas.sort((a: any, b: any) => 
-        new Date(b.paymentDate || b.clientPaymentDate || b.dueDate).getTime() - new Date(a.paymentDate || a.clientPaymentDate || a.dueDate).getTime()
-      )[0];
-
-      // Verificar se cliente tem período ativo baseado na última cobrança paga
-      if (ultimaCobrancaPaga) {
-        const dataPagamento = new Date(ultimaCobrancaPaga.paymentDate || ultimaCobrancaPaga.clientPaymentDate || ultimaCobrancaPaga.dueDate);
-        
-        // Verificar se existe próxima cobrança registrada
-        const proximaCobranca = cobrancas.find((c: any) => 
-          (c.status !== 'RECEIVED' && c.status !== 'CONFIRMED') && new Date(c.dueDate) > dataPagamento
-        );
-        
-        let dataVencimentoAssinatura: Date;
-        
-        if (proximaCobranca) {
-          // Se tem próxima cobrança, usar a data de vencimento dela
-          dataVencimentoAssinatura = new Date(proximaCobranca.dueDate);
-        } else {
-          // Se não tem próxima cobrança, considerar 30 dias a partir do pagamento
-          dataVencimentoAssinatura = new Date(dataPagamento);
-          dataVencimentoAssinatura.setDate(dataVencimentoAssinatura.getDate() + 30);
-        }
-        
-        // Se ainda está dentro do período válido = ativo
-        if (hoje <= dataVencimentoAssinatura) {
-          console.log(`Cliente ${cliente.id} ativo: dentro do período válido da assinatura`);
-          return 'ativo';
-        }
-      }
-
-      // Verificar se tem cobrança pendente vencida
-      const cobrancasPendentes = cobrancas.filter((c: any) => c.status !== 'RECEIVED' && c.status !== 'CONFIRMED');
-      const cobrancasVencidasNaoPagas = cobrancasPendentes.filter((cobranca: any) => {
-        const dataVencimento = new Date(cobranca.dueDate);
-        dataVencimento.setHours(23, 59, 59, 999);
-        return dataVencimento < hoje;
-      });
-
-      // Se tem cobrança vencida não paga = inadimplente
-      if (cobrancasVencidasNaoPagas.length > 0) {
-        console.log(`Cliente ${cliente.id} inadimplente: ${cobrancasVencidasNaoPagas.length} cobrança(s) vencida(s)`);
-        return 'inadimplente';
-      }
-
-      // Verificar se tem cobrança pendente dentro do prazo
-      const cobrancasPendentesNoPrazo = cobrancasPendentes.filter((cobranca: any) => {
-        const dataVencimento = new Date(cobranca.dueDate);
-        dataVencimento.setHours(23, 59, 59, 999);
-        return dataVencimento >= hoje;
-      });
-
-      // Se tem cobrança pendente no prazo = ativo
-      if (cobrancasPendentesNoPrazo.length > 0) {
-        console.log(`Cliente ${cliente.id} ativo: cobrança pendente ainda no prazo`);
-        return 'ativo';
-      }
-
-      // Por padrão, considera ativo (clientes novos ou sem cobranças)
-      console.log(`Cliente ${cliente.id} ativo: cliente novo ou sem cobranças pendentes`);
-      return 'ativo';
-      
-    } catch (error) {
-      console.error(`Erro ao verificar status do cliente ${cliente.id}:`, error);
-      return 'ativo'; // Default para ativo em caso de erro
-    }
-  }
-  
-  // =====================================================
-  // ROTA UNIFICADA CLIENTES ASAAS POR STATUS
-  // =====================================================
-
-  app.get('/api/clientes/unificados-status', async (req: Request, res: Response) => {
-    try {
-      const mesFiltro = req.query.mes as string || new Date().toISOString().slice(0, 7); // YYYY-MM
-      console.log(`📅 Filtrando clientes por mês: ${mesFiltro}`);
-      
-      const asaasTrato = process.env.ASAAS_TRATO;
-      const asaasAndrey = process.env.ASAAS_AND;
-      
-      const clientesAtivos: any[] = [];
-      const clientesAtrasados: any[] = [];
-      
-      // Buscar clientes da conta ASAAS_TRATO
-      if (asaasTrato) {
-        try {
-          const response = await fetch('https://www.asaas.com/api/v3/customers?limit=100', {
-            headers: {
-              'access_token': asaasTrato,
-              'Content-Type': 'application/json'
-            }
-          });
-          
-          if (response.ok) {
-            const data = await response.json();
-            data.data.forEach((cliente: any) => {
-              const clienteFormatado = {
-                ...cliente,
-                conta: 'ASAAS_TRATO'
-              };
-              
-              const status = determinarStatusClientePorMes(clienteFormatado, mesFiltro);
-              if (status === 'ativo') {
-                clientesAtivos.push({...clienteFormatado, status: 'ativo'});
-              } else {
-                clientesAtrasados.push({...clienteFormatado, status: 'atrasado'});
-              }
-            });
-          }
-        } catch (error) {
-          console.error('Erro ao buscar clientes ASAAS_TRATO:', error);
-        }
-      }
-      
-      // Buscar clientes da conta ASAAS_AND
-      if (asaasAndrey) {
-        try {
-          const response = await fetch('https://www.asaas.com/api/v3/customers?limit=100', {
-            headers: {
-              'access_token': asaasAndrey,
-              'Content-Type': 'application/json'
-            }
-          });
-          
-          if (response.ok) {
-            const data = await response.json();
-            data.data.forEach((cliente: any) => {
-              const clienteFormatado = {
-                ...cliente,
-                conta: 'ASAAS_AND'
-              };
-              
-              const status = determinarStatusClientePorMes(clienteFormatado, mesFiltro);
-              if (status === 'ativo') {
-                clientesAtivos.push({...clienteFormatado, status: 'ativo'});
-              } else {
-                clientesAtrasados.push({...clienteFormatado, status: 'atrasado'});
-              }
-            });
-          }
-        } catch (error) {
-          console.error('Erro ao buscar clientes ASAAS_AND:', error);
-        }
-      }
-      
-      // Organizar clientes por status baseado em cobranças
-      const ativos: any[] = [];
-      const inadimplentes: any[] = [];
-      const aguardandoPagamento: any[] = [];
-
-      // Buscar clientes externos e adicioná-los como ativos
-      let clientesExternos: any[] = [];
-      try {
-        const clientesDB = await db.select().from(schema.clientes)
-          .where(eq(schema.clientes.origem, 'EXTERNO'));
-
-        clientesExternos = clientesDB.map(cliente => ({
-          id: cliente.id,
-          nome: cliente.nome,
-          email: cliente.email,
-          telefone: cliente.telefone,
-          valor: parseFloat(cliente.planoValor),
-          plano: cliente.planoNome,
-          dataVencimento: cliente.dataVencimentoAssinatura,
-          dataInicio: cliente.dataInicioAssinatura,
-          status: 'ativo',
-          conta: 'PAGAMENTO_EXTERNO',
-          formaPagamento: cliente.formaPagamento
-        }));
-
-        console.log(`📊 PAGAMENTO_EXTERNO: ${clientesExternos.length} clientes encontrados`);
-      } catch (error) {
-        console.error('Erro ao buscar clientes externos:', error);
-      }
-
-      // Combinar todos os clientes das duas contas + externos e eliminar duplicatas
-      const todosClientesTemp = [...clientesAtivos, ...clientesAtrasados, ...clientesExternos];
-      
-      // Eliminar duplicatas usando email como chave única
-      const clientesUnicosMap = new Map();
-      todosClientesTemp.forEach(cliente => {
-        const chave = cliente.email || cliente.nome;
-        if (!clientesUnicosMap.has(chave)) {
-          clientesUnicosMap.set(chave, cliente);
-        } else {
-          // Manter o cliente com valor maior (assinatura mais recente/cara)
-          const clienteExistente = clientesUnicosMap.get(chave);
-          if (cliente.valor > clienteExistente.valor) {
-            clientesUnicosMap.set(chave, cliente);
-          }
-        }
-      });
-      
-      const todosClientes = Array.from(clientesUnicosMap.values());
-      
-      console.log(`🔍 Analisando ${todosClientes.length} clientes únicos para verificar status baseado em cobranças...`);
-      
-      // Processar cada cliente para verificar status baseado em cobranças
-      for (const cliente of todosClientes) {
-        try {
-          // Clientes externos são sempre ativos
-          if (cliente.conta === 'PAGAMENTO_EXTERNO') {
-            ativos.push({...cliente, status: 'ativo', origem: 'Pagamento Externo'});
-            continue;
-          }
-
-          const apiKey = cliente.conta === 'ASAAS_TRATO' ? asaasTrato : asaasAndrey;
-          if (apiKey) {
-            const status = await verificarStatusCliente(cliente, apiKey);
-            
-            if (status === 'inadimplente') {
-              inadimplentes.push({...cliente, status: 'inadimplente'});
-            } else {
-              ativos.push({...cliente, status: 'ativo'});
-            }
-          } else {
-            // Se não tem API key, considera ativo por padrão
-            ativos.push({...cliente, status: 'ativo'});
-          }
-        } catch (error) {
-          console.error(`Erro ao processar cliente ${cliente.id}:`, error);
-          // Em caso de erro, considera ativo
-          ativos.push({...cliente, status: 'ativo'});
-        }
-      }
-
-      console.log(`✅ Análise concluída: ${ativos.length} ativos, ${inadimplentes.length} inadimplentes`);
-
-      res.json({
-        success: true,
-        total: ativos.length + inadimplentes.length,
-        ativos: {
-          total: ativos.length,
-          clientes: ativos
-        },
-        inativos: {
-          total: inadimplentes.length,
-          clientes: inadimplentes
-        },
-        aguardandoPagamento: {
-          total: 0,
-          clientes: []
-        },
-        timestamp: new Date().toISOString()
-      });
-    } catch (error) {
-      console.error('Erro na rota clientes unificados:', error);
-      res.status(500).json({ 
-        success: false,
-        error: error instanceof Error ? error.message : 'Erro desconhecido'
-      });
-    }
-  });
-
-  // =====================================================
-  // API GESTÃO DE ASSINATURAS E CLIENTES
-  // =====================================================
-
-  // POST - Cadastro manual de cliente
-  app.post('/api/clientes/cadastro-manual', async (req: Request, res: Response) => {
-    try {
-      const { nome, telefone, email } = req.body;
-
-      // Validações básicas
-      if (!nome || !telefone || !email) {
-        return res.status(400).json({
-          success: false,
-          message: 'Nome, telefone e email são obrigatórios'
-        });
-      }
-
-      // Verificar se cliente já existe (por email ou telefone)
-      const clienteExistente = await db.select()
-        .from(schema.clientes)
-        .where(
-          sql`${schema.clientes.email} = ${email} OR ${schema.clientes.telefone} = ${telefone}`
-        )
-        .limit(1);
-
-      let clienteId: number;
-      let isUpdate = false;
-
-      if (clienteExistente.length > 0) {
-        // Cliente existe - atualizar dados
-        const cliente = clienteExistente[0];
-        
-        await db.update(schema.clientes)
-          .set({
-            nome,
-            telefone,
-            email
-          })
-          .where(eq(schema.clientes.id, cliente.id));
-
-        clienteId = cliente.id;
-        isUpdate = true;
-        
-        console.log(`🔄 Cliente atualizado: ${nome} (ID: ${clienteId})`);
-      } else {
-        // Cliente novo - inserir
-        const novoCliente = await db.insert(schema.clientes)
-          .values({
-            nome,
-            telefone,
-            email,
-            origem: 'EXTERNO',
-            planoNome: 'Cadastro Manual',
-            planoValor: '0.00',
-            formaPagamento: 'PENDENTE',
-            statusAssinatura: 'INATIVO',
-            dataInicioAssinatura: new Date(),
-            dataVencimentoAssinatura: new Date()
-          })
-          .returning({ id: schema.clientes.id });
-
-        clienteId = novoCliente[0].id;
-        
-        console.log(`✅ Novo cliente cadastrado: ${nome} (ID: ${clienteId})`);
-      }
-
-      res.json({
-        success: true,
-        clienteId,
-        isUpdate,
-        message: isUpdate 
-          ? 'Dados do cliente atualizados com sucesso'
-          : 'Cliente cadastrado com sucesso'
-      });
-
-    } catch (error) {
-      console.error('Erro ao cadastrar cliente:', error);
-      res.status(500).json({
-        success: false,
-        message: 'Erro interno do servidor',
-        error: error instanceof Error ? error.message : 'Erro desconhecido'
-      });
-    }
-  });
-
-  // POST - Importação em lote de clientes via Excel
-  app.post('/api/clientes/importar-lote', async (req: Request, res: Response) => {
-    try {
-      const { clientes } = req.body;
-
-      if (!clientes || !Array.isArray(clientes) || clientes.length === 0) {
-        return res.status(400).json({
-          success: false,
-          message: 'Lista de clientes é obrigatória'
-        });
-      }
-
-      let novos = 0;
-      let atualizados = 0;
-      const erros: { linha: number; motivo: string }[] = [];
-
-      console.log(`📥 Iniciando importação em lote de ${clientes.length} clientes...`);
-
-      // Processar cada cliente
-      for (let index = 0; index < clientes.length; index++) {
-        const cliente = clientes[index];
-        const linha = index + 2; // +2 porque começa da linha 2 (linha 1 é cabeçalho)
-
-        try {
-          const { nome, telefone, email } = cliente;
-
-          // Validações básicas
-          if (!nome || nome.trim().length < 2) {
-            erros.push({ linha, motivo: "Nome deve ter pelo menos 2 caracteres" });
-            continue;
-          }
-
-          if (!telefone || telefone.trim().length < 10) {
-            erros.push({ linha, motivo: "Telefone deve ter pelo menos 10 dígitos" });
-            continue;
-          }
-
-          // Email é opcional, mas se fornecido deve ser válido
-          let emailValidado = null;
-          if (email && email.trim() !== "") {
-            const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-            if (!emailRegex.test(email.trim())) {
-              erros.push({ linha, motivo: "Email inválido" });
-              continue;
-            }
-            emailValidado = email.trim();
-          }
-
-          // Verificar se cliente já existe (por telefone ou email)
-          let whereCondition;
-          if (emailValidado) {
-            whereCondition = sql`${schema.clientes.email} = ${emailValidado} OR ${schema.clientes.telefone} = ${telefone.trim()}`;
-          } else {
-            whereCondition = sql`${schema.clientes.telefone} = ${telefone.trim()}`;
-          }
-
-          const clienteExistente = await db.select()
-            .from(schema.clientes)
-            .where(whereCondition)
-            .limit(1);
-
-          if (clienteExistente.length > 0) {
-            // Cliente existe - atualizar dados
-            const clienteAtual = clienteExistente[0];
-            
-            const updateData: any = {
-              nome: nome.trim(),
-              telefone: telefone.trim()
-            };
-
-            if (emailValidado) {
-              updateData.email = emailValidado;
-            }
-            
-            await db.update(schema.clientes)
-              .set(updateData)
-              .where(eq(schema.clientes.id, clienteAtual.id));
-
-            atualizados++;
-            console.log(`🔄 Cliente atualizado via Excel: ${nome.trim()}`);
-          } else {
-            // Cliente novo - inserir
-            const insertData: any = {
-              nome: nome.trim(),
-              telefone: telefone.trim(),
-              email: emailValidado || 'sem-email@exemplo.com',
-              origem: 'EXTERNO',
-              planoNome: 'Importação Excel',
-              planoValor: '0.00',
-              formaPagamento: 'PENDENTE',
-              statusAssinatura: 'INATIVO',
-              dataInicioAssinatura: new Date(),
-              dataVencimentoAssinatura: new Date()
-            };
-
-            await db.insert(schema.clientes).values(insertData);
-
-            novos++;
-            console.log(`✅ Novo cliente via Excel: ${nome.trim()}`);
-          }
-
-        } catch (error) {
-          console.error(`Erro ao processar cliente linha ${linha}:`, error);
-          erros.push({ 
-            linha, 
-            motivo: `Erro no processamento: ${error instanceof Error ? error.message : 'Erro desconhecido'}` 
-          });
-        }
-      }
-
-      console.log(`📊 Importação concluída: ${novos} novos, ${atualizados} atualizados, ${erros.length} erros`);
-
-      res.json({
-        success: true,
-        novos,
-        atualizados,
-        erros,
-        message: `Importação concluída: ${novos} novos clientes, ${atualizados} atualizados`
-      });
-
-    } catch (error) {
-      console.error('Erro na importação em lote:', error);
-      res.status(500).json({
-        success: false,
-        message: 'Erro interno do servidor na importação',
-        error: error instanceof Error ? error.message : 'Erro desconhecido'
-      });
-    }
-  });
-
-  // Endpoint para buscar clientes pagantes do mês vigente
-  app.get('/api/clientes/pagamentos-mes', async (req: Request, res: Response) => {
     try {
       const asaasTrato = process.env.ASAAS_TRATO;
       const asaasAndrey = process.env.ASAAS_AND;
@@ -1393,70 +874,6 @@ export async function registerRoutes(app: Express): Promise<Express> {
     }
   });
 
-  // =====================================================
-  // CRIAR CLIENTE NO ASAAS (APENAS PRODUÇÃO)
-  // =====================================================
-
-  app.post("/api/asaas/criar-cliente", async (req: Request, res: Response) => {
-    try {
-      const { nome, email, telefone, cpf } = req.body;
-      
-      if (!nome) {
-        return res.status(400).json({ message: 'Nome é obrigatório' });
-      }
-
-      const asaasApiKey = process.env.ASAAS_AND;
-      if (!asaasApiKey) {
-        return res.status(500).json({ message: 'Chave API ASAAS não configurada' });
-      }
-
-      const customerData = {
-        name: nome,
-        email: email || undefined,
-        phone: telefone || undefined,
-        cpfCnpj: cpf || undefined
-      };
-
-      // PRODUÇÃO APENAS: URL fixa para produção
-      const baseUrl = 'https://www.asaas.com/api/v3';
-
-      const response = await fetch(`${baseUrl}/customers`, {
-        method: 'POST',
-        headers: {
-          'access_token': asaasApiKey,
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(customerData)
-      });
-
-      const result = await response.json();
-
-      if (response.ok) {
-        res.json({
-          success: true,
-          customer: result,
-          ambiente: 'PRODUÇÃO',
-          baseUrl: baseUrl,
-          message: 'Cliente criado com sucesso'
-        });
-      } else {
-        res.status(400).json({
-          success: false,
-          message: 'Erro ao criar cliente',
-          ambiente: 'PRODUÇÃO',
-          baseUrl: baseUrl,
-          error: result
-        });
-      }
-    } catch (error) {
-      console.error('Erro ao criar cliente:', error);
-      res.status(500).json({ 
-        success: false,
-        message: 'Erro interno do servidor',
-        error: error instanceof Error ? error.message : 'Erro desconhecido'
-      });
-    }
-  });
 
   // =====================================================
   // CRIAR PLANO DE ASSINATURA (ENDPOINT GERENCIAR-ASSINATURAS)
@@ -1570,72 +987,6 @@ export async function registerRoutes(app: Express): Promise<Express> {
     }
   });
 
-  // =====================================================
-  // CRIAR ASSINATURA NO ASAAS (APENAS PRODUÇÃO)
-  // =====================================================
-
-  app.post("/api/asaas/criar-assinatura", async (req: Request, res: Response) => {
-    try {
-      const { customerId, value, description } = req.body;
-      
-      if (!customerId || !value) {
-        return res.status(400).json({ message: 'customerId e value são obrigatórios' });
-      }
-
-      const asaasApiKey = process.env.ASAAS_AND;
-      if (!asaasApiKey) {
-        return res.status(500).json({ message: 'Chave API ASAAS não configurada' });
-      }
-
-      const subscriptionData = {
-        customer: customerId,
-        billingType: 'PIX',
-        nextDueDate: new Date().toISOString().split('T')[0],
-        value: value,
-        cycle: 'MONTHLY',
-        description: description || 'Assinatura Mensal'
-      };
-
-      // PRODUÇÃO APENAS: URL fixa para produção
-      const baseUrl = 'https://www.asaas.com/api/v3';
-
-      const response = await fetch(`${baseUrl}/subscriptions`, {
-        method: 'POST',
-        headers: {
-          'access_token': asaasApiKey,
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(subscriptionData)
-      });
-
-      const result = await response.json();
-
-      if (response.ok) {
-        res.json({
-          success: true,
-          subscription: result,
-          ambiente: 'PRODUÇÃO',
-          baseUrl: baseUrl,
-          message: 'Assinatura criada com sucesso'
-        });
-      } else {
-        res.status(400).json({
-          success: false,
-          message: 'Erro ao criar assinatura',
-          ambiente: 'PRODUÇÃO',
-          baseUrl: baseUrl,
-          error: result
-        });
-      }
-    } catch (error) {
-      console.error('Erro ao criar assinatura:', error);
-      res.status(500).json({ 
-        success: false,
-        message: 'Erro interno do servidor',
-        error: error instanceof Error ? error.message : 'Erro desconhecido'
-      });
-    }
-  });
 
   // =====================================================
   // WEBHOOK ASAAS (APENAS PRODUÇÃO)
@@ -2594,100 +1945,7 @@ export async function registerRoutes(app: Express): Promise<Express> {
     }
   });
 
-  // =====================================================
-  // AUTENTICAÇÃO BÁSICA
-  // =====================================================
 
-  // Login com autenticação real usando cookies
-  app.post('/api/auth/login', async (req: Request, res: Response) => {
-    try {
-      const { email, password, rememberMe } = req.body;
-      
-      if (!email || !password) {
-        return res.status(400).json({ message: 'Email e senha são obrigatórios' });
-      }
-      
-      // Buscar usuário no banco
-      const [user] = await db.select().from(schema.users).where(eq(schema.users.email, email));
-      
-      if (!user) {
-        return res.status(401).json({ message: 'Email ou senha incorretos' });
-      }
-      
-      // Verificar senha
-      const senhaValida = await bcrypt.compare(password, user.password);
-      if (!senhaValida) {
-        return res.status(401).json({ message: 'Email ou senha incorretos' });
-      }
-      
-      // Configurar duração da sessão baseado em rememberMe
-      const sessionDuration = rememberMe 
-        ? 30 * 24 * 60 * 60 * 1000 // 30 dias se "Me manter logado" estiver marcado
-        : 24 * 60 * 60 * 1000;     // 24 horas para sessão normal
-      
-      // Configurar cookies de autenticação
-      const cookieOptions = {
-        httpOnly: true,
-        secure: process.env.NODE_ENV === 'production',
-        sameSite: 'lax' as const,
-        maxAge: sessionDuration
-      };
-      
-      res.cookie('user_id', user.id.toString(), cookieOptions);
-      res.cookie('user_email', user.email, cookieOptions);
-      res.cookie('user_role', user.role, cookieOptions);
-      res.cookie('remember_me', rememberMe ? 'true' : 'false', cookieOptions);
-      
-      res.json({
-        success: true,
-        user: {
-          id: user.id,
-          email: user.email,
-          role: user.role,
-          nome: user.nome
-        }
-      });
-    } catch (error) {
-      console.error('Erro no login:', error);
-      res.status(500).json({ message: 'Erro interno do servidor' });
-    }
-  });
-
-  // Logout com limpeza completa dos cookies
-  app.post('/api/auth/logout', (req: Request, res: Response) => {
-    res.clearCookie('user_id');
-    res.clearCookie('user_email');
-    res.clearCookie('user_role');
-    res.json({ success: true, message: 'Logout realizado com sucesso' });
-  });
-
-  // Dados do usuário com verificação via cookies
-  app.get('/api/auth/me', async (req: Request, res: Response) => {
-    try {
-      const userId = req.cookies.user_id;
-      
-      if (!userId) {
-        return res.status(401).json({ message: 'Não autenticado' });
-      }
-      
-      // Buscar dados atualizados do usuário
-      const [user] = await db.select().from(schema.users).where(eq(schema.users.id, parseInt(userId)));
-      
-      if (!user) {
-        return res.status(401).json({ message: 'Usuário não encontrado' });
-      }
-      
-      res.json({
-        id: user.id,
-        email: user.email,
-        role: user.role,
-        nome: user.nome
-      });
-    } catch (error) {
-      console.error('Erro ao buscar dados do usuário:', error);
-      res.status(500).json({ message: 'Erro interno do servidor' });
-    }
-  });
 
   // =====================================================
   // SISTEMA DE PROFISSIONAIS (BARBEIROS E RECEPCIONISTAS)
